Read the file to be uploaded
Begin the upload script by testing to make sure we can get the filename
and the some data back from fs.readFile.

Use an empty string as a default arg for the filepath so that if
we don't pass an arg when running the script we get a sensible
error message (file not found).