Implement upload using AWS.S3.upload() method
use dotenv for secret keys
use aws-sdk for the upload
configure AWS.S3 to use our credentials
do a test upload
change the log

Next, we need to generate better names for our files. These need to be
unique with a low change of duplication, so we pseduo random number
generator.